moved some components into coursesPage component
