+++ src/app/app.html
@@ -1,5 +1,3 @@
-<app-header></app-header>
+<courses-page></courses-page>
 
-<action-bar></action-bar>
-<course-list></course-list>
 <app-footer></app-footer>
