+++ src/app/components/coursesPage/actionBar/search/index.ts
