+++ src/app/components/coursesPage/actionBar/search/search.html
