+++ src/app/components/coursesPage/courseList/course/course.ts
