+++ src/app/components/coursesPage/courseList/course/index.ts
