+++ src/app/components/coursesPage/courseList/courseList.ts
