+++ src/app/components/coursesPage/courseList/index.ts
