+++ src/app/components/coursesPage/coursesPage.html
@@ -1,0 +1,4 @@
+<app-header></app-header>
+
+<action-bar></action-bar>
+<course-list></course-list>
