+++ src/app/components/coursesPage/coursesPage.ts
@@ -1,0 +1,12 @@
+import {Component} from '@angular/core'
+
+@Component({
+  selector: 'courses-page',
+  providers: [],
+  templateUrl: './coursesPage.html', 
+  styleUrls: []
+})
+export class CoursesPageComponent {
+  constructor() {
+  }
+}
