+++ src/app/components/coursesPage/index.ts
@@ -1,0 +1,2 @@
+export * from './coursesPage';
+console.log('Courses Page component');
